+++ eslint.config.mjs
@@ -12,6 +12,6 @@ export default tseslint.config(
     },
   },
   {
-    ignores: ["node_modules/", "dist/"],
+    ignores: ["node_modules/", "dist/", "eval/"],
   },
 );